putting spoiler log back together
