+++ rust/maprando/src/bin/maprando-test.rs
@@ -8,13 +8,14 @@ use maprando::patch::Rom;
 use maprando::patch::make_rom;
 use maprando::preset::PresetData;
 use maprando::randomize::{
-    Randomization, Randomizer, SpoilerLog, get_difficulty_tiers, get_objectives, order_map_areas,
+    Randomization, Randomizer, get_difficulty_tiers, get_objectives, order_map_areas,
     randomize_doors, randomize_map_areas,
 };
 use maprando::settings::{
     AreaAssignment, ItemProgressionSettings, QualityOfLifeSettings, RandomizerSettings,
     SkillAssumptionSettings, StartLocationMode,
 };
+use maprando::spoiler_log::SpoilerLog;
 use maprando::spoiler_map;
 use maprando_game::{GameData, Map};
 use rand::{RngCore, SeedableRng};
