+++ rust/maprando/src/lib.rs
@@ -11,5 +11,6 @@ pub mod preset;
 pub mod randomize;
 pub mod seed_repository;
 pub mod settings;
+pub mod spoiler_log;
 pub mod spoiler_map;
 pub mod traverse;
