+++ rust/maprando/src/randomize.rs
@@ -9,6 +9,10 @@ use crate::settings::{
     MotherBrainFight, Objective, ObjectiveSetting, ProgressionRate, RandomizerSettings,
     SaveAnimals, SkillAssumptionSettings, StartLocationMode, WallJump,
 };
+use crate::spoiler_log::{
+    SpoilerDetails, SpoilerLog, SpoilerRoomLoc, SpoilerRouteEntry, SpoilerStartLocation,
+    SpoilerSummary, get_spoiler_game_data, get_spoiler_log, get_spoiler_route,
+};
 use crate::traverse::{
     LockedDoorData, NUM_COST_METRICS, Traverser, apply_requirement, get_bireachable_idxs,
     get_one_way_reachable_idx, get_spoiler_trail_ids,
@@ -317,15 +321,6 @@ pub struct Randomization {
 struct SelectItemsOutput {
     key_items: Vec<Item>,
     other_items: Vec<Item>,
-}
-
-struct VertexInfo {
-    area_name: String,
-    room_id: usize,
-    room_name: String,
-    room_coords: (usize, usize),
-    node_name: String,
-    node_id: usize,
 }
 
 pub struct StartLocationData {
@@ -4243,271 +4238,14 @@ impl<'r> Randomizer<'r> {
         EssentialSpoilerData { item_spoiler_info }
     }
 
-    fn get_spoiler_game_data(&self) -> SpoilerGameData {
-        let mut rooms: Vec<SpoilerRoom> = vec![];
-        let mut nodes: Vec<SpoilerNode> = vec![];
-        for &room_ptr in &self.game_data.room_ptrs {
-            let room_id = self.game_data.raw_room_id_by_ptr[&room_ptr];
-            let room = &self.game_data.room_json_map[&room_id];
-            rooms.push(SpoilerRoom {
-                room_id,
-                name: room["name"].as_str().unwrap().to_string(),
-            });
-
-            for node in room["nodes"].members() {
-                let node_id = node["id"].as_usize().unwrap();
-                let node_name = node["name"].as_str().unwrap().to_string();
-                nodes.push(SpoilerNode {
-                    room_id,
-                    node_id,
-                    name: node_name,
-                });
-            }
-        }
-
-        let mut links: Vec<SpoilerLink> = vec![];
-        for link in self
-            .base_links_data
-            .links
-            .iter()
-            .chain(self.seed_links_data.links.iter())
-        {
-            links.push(SpoilerLink {
-                from_vertex_id: link.from_vertex_id,
-                to_vertex_id: link.to_vertex_id,
-                strat_id: link.strat_id,
-                strat_name: link.strat_name.clone(),
-            });
-        }
-        SpoilerGameData {
-            rooms,
-            nodes,
-            vertices: self.game_data.vertex_isv.keys.clone(),
-            links,
-        }
-    }
-
     pub fn get_randomization<R: Rng>(
         &self,
         state: &RandomizationState,
-        spoiler_summaries: Vec<SpoilerSummary>,
-        spoiler_details: Vec<SpoilerDetails>,
         seed: usize,
         display_seed: usize,
         rng: &mut R,
-        traverser_pair: &mut TraverserPair
+        traverser_pair: &mut TraverserPair,
     ) -> Result<(Randomization, SpoilerLog)> {
-        // Compute the first step on which each node becomes reachable/bireachable:
-        let mut node_reachable_step: HashMap<(RoomId, NodeId), usize> = HashMap::new();
-        let mut node_bireachable_step: HashMap<(RoomId, NodeId), usize> = HashMap::new();
-        let mut map_tile_reachable_step: HashMap<(RoomId, (usize, usize)), usize> = HashMap::new();
-        let mut map_tile_bireachable_step: HashMap<(RoomId, (usize, usize)), usize> =
-            HashMap::new();
-
-
-
-
-                            // spoiler_flag_summaries.push(self.get_spoiler_flag_summary(
-                    //     state,
-                    //     flag_vertex_id,
-                    //     flag_id,
-                    // ));
-                    // spoiler_flag_details.push(self.get_spoiler_flag_details_one_way(
-                    //     state,
-                    //     flag_vertex_id,
-                    //     flag_id,
-                    //     i,
-                    //     &traverser_pair.forward,
-                    // ));
-
-                    // spoiler_flag_summaries.push(self.get_spoiler_flag_summary(
-                    //     state,
-                    //     flag_vertex_id,
-                    //     flag_id,
-                    // ));
-                    // spoiler_flag_details.push(self.get_spoiler_flag_details(
-                    //     state,
-                    //     flag_vertex_id,
-                    //     flag_id,
-                    //     i,
-                    //     traverser_pair,
-                    // ));
-
-                    // spoiler_door_summaries.push(self.get_spoiler_door_summary(door_vertex_id, i));
-                    // spoiler_door_details.push(self.get_spoiler_door_details(
-                    //     state,
-                    //     door_vertex_id,
-                    //     i,
-                    //     traverser_pair,
-                    // ));
-            let spoiler_summary = self.get_spoiler_summary(
-                &orig_global_state,
-                state,
-                state,
-                spoiler_flag_summaries,
-                spoiler_door_summaries,
-            );
-            let spoiler_details = self.get_spoiler_details(
-                &orig_global_state,
-                state,
-                state,
-                spoiler_flag_details,
-                spoiler_door_details,
-                traverser_pair,
-            );
-
-
-
-        // for (step, debug_data) in debug_data_vec.iter_mut().enumerate() {
-        //     for (
-        //         v,
-        //         VertexKey {
-        //             room_id, node_id, ..
-        //         },
-        //     ) in self.game_data.vertex_isv.keys.iter().enumerate()
-        //     {
-        //         if node_bireachable_step.contains_key(&(*room_id, *node_id)) {
-        //             continue;
-        //         }
-        //         if get_bireachable_idxs(
-        //             &debug_data.global_state,
-        //             v,
-        //             &debug_data.forward,
-        //             &debug_data.reverse,
-        //         )
-        //         .is_some()
-        //         {
-        //             node_bireachable_step.insert((*room_id, *node_id), step);
-        //             let room_ptr = self.game_data.room_ptr_by_id[room_id];
-        //             let room_idx = self.game_data.room_idx_by_ptr[&room_ptr];
-        //             if let Some(coords) = self.game_data.node_tile_coords.get(&(*room_id, *node_id))
-        //             {
-        //                 for (x, y) in coords.iter().copied() {
-        //                     let key = if *room_id == 322 {
-        //                         // Adjust for East Pants Room being offset by one screen right and down from Pants Room
-        //                         (room_idx, (x + 1, y + 1))
-        //                     } else if *room_id == 313 {
-        //                         // Adjust Homing Geemer Room being offset from West Ocean:
-        //                         (room_idx, (x + 5, y + 2))
-        //                     } else {
-        //                         (room_idx, (x, y))
-        //                     };
-        //                     if !map_tile_bireachable_step.contains_key(&key) {
-        //                         map_tile_bireachable_step.insert(key, step);
-        //                     }
-        //                 }
-        //             }
-        //         }
-
-        //         if node_reachable_step.contains_key(&(*room_id, *node_id)) {
-        //             continue;
-        //         }
-        //         if debug_data.forward.cost[v]
-        //             .iter()
-        //             .any(|&x| f32::is_finite(x))
-        //         {
-        //             node_reachable_step.insert((*room_id, *node_id), step);
-        //             let room_ptr = self.game_data.room_ptr_by_id[room_id];
-        //             let room_idx = self.game_data.room_idx_by_ptr[&room_ptr];
-        //             if let Some(coords) = self.game_data.node_tile_coords.get(&(*room_id, *node_id))
-        //             {
-        //                 for (x, y) in coords.iter().copied() {
-        //                     let key = if *room_id == 322 {
-        //                         // Adjust for East Pants Room being offset by one screen right and down from Pants Room
-        //                         (room_idx, (x + 1, y + 1))
-        //                     } else if *room_id == 313 {
-        //                         // Adjust Homing Geemer Room being offset from West Ocean:
-        //                         (room_idx, (x + 5, y + 2))
-        //                     } else {
-        //                         (room_idx, (x, y))
-        //                     };
-        //                     if !map_tile_reachable_step.contains_key(&key) {
-        //                         map_tile_reachable_step.insert(key, step);
-        //                     }
-        //                 }
-        //             }
-        //         }
-        //     }
-        // }
-
-        let item_placement: Vec<Item> = state
-            .item_location_state
-            .iter()
-            .map(|x| x.placed_item.unwrap())
-            .collect();
-        let spoiler_all_items = state
-            .item_location_state
-            .iter()
-            .enumerate()
-            .map(|(i, x)| {
-                let (r, n) = self.game_data.item_locations[i];
-                let item_vertex_info = self.get_vertex_info_by_id(r, n);
-                let room_id = item_vertex_info.room_id;
-                let node_id = item_vertex_info.node_id;
-                let node_coords = self.game_data.node_coords[&(room_id, node_id)];
-                let coords = (
-                    item_vertex_info.room_coords.0 + node_coords.0,
-                    item_vertex_info.room_coords.1 + node_coords.1,
-                );
-                let location = SpoilerLocation {
-                    area: item_vertex_info.area_name,
-                    room_id,
-                    room: item_vertex_info.room_name,
-                    node_id,
-                    node: item_vertex_info.node_name,
-                    coords,
-                };
-                let item = x.placed_item.unwrap();
-                SpoilerItemLoc {
-                    item: Item::VARIANTS[item as usize].to_string(),
-                    location,
-                }
-            })
-            .collect();
-
-        let mut spoiler_all_rooms: Vec<SpoilerRoomLoc> = Vec::new();
-        for (room_idx, room_coords) in self.map.rooms.iter().enumerate() {
-            if !self.map.room_mask[room_idx] {
-                continue;
-            }
-            let room_geom = &self.game_data.room_geometry[room_idx];
-            let room_id = self.game_data.room_id_by_ptr[&room_geom.rom_address];
-            let room = self.game_data.room_json_map[&room_id]["name"]
-                .as_str()
-                .unwrap()
-                .to_string();
-            let map = if room_idx == self.game_data.toilet_room_idx {
-                vec![vec![1; 1]; 10]
-            } else {
-                room_geom.map.clone()
-            };
-            let height = map.len();
-            let width = map[0].len();
-            let mut map_reachable_step: Vec<Vec<u8>> = vec![vec![255; width]; height];
-            let mut map_bireachable_step: Vec<Vec<u8>> = vec![vec![255; width]; height];
-            for y in 0..height {
-                for x in 0..width {
-                    if map[y][x] != 0 {
-                        let key = (room_idx, (x, y));
-                        if let Some(step) = map_tile_reachable_step.get(&key) {
-                            map_reachable_step[y][x] = *step as u8;
-                        }
-                        if let Some(step) = map_tile_bireachable_step.get(&key) {
-                            map_bireachable_step[y][x] = *step as u8;
-                        }
-                    }
-                }
-            }
-            spoiler_all_rooms.push(SpoilerRoomLoc {
-                room_id,
-                room,
-                map,
-                map_reachable_step,
-                map_bireachable_step,
-                coords: *room_coords,
-            });
-        }
-
         let save_animals = if self.settings.save_animals == SaveAnimals::Random {
             if rng.gen_bool(0.5) {
                 SaveAnimals::Yes
@@ -4518,49 +4256,13 @@ impl<'r> Randomizer<'r> {
             self.settings.save_animals
         };
 
-        let spoiler_escape = escape_timer::compute_escape_data(
-            self.game_data,
-            self.map,
-            self.settings,
-            save_animals != SaveAnimals::No,
-            &self.difficulty_tiers[0],
-        )?;
+        let spoiler_log = get_spoiler_log(self, state, traverser_pair, save_animals)?;
 
-        let spoiler_objectives: Vec<String> = self
-            .objectives
+        let item_placement: Vec<Item> = state
+            .item_location_state
             .iter()
-            .map(|x| x.get_flag_name().to_owned())
+            .map(|x| x.placed_item.unwrap())
             .collect();
-
-        let hub_room_id = state.hub_location.room_id;
-        let hub_room_name = self.game_data.room_json_map[&hub_room_id]["name"]
-            .as_str()
-            .unwrap()
-            .to_string();
-        let spoiler_log = SpoilerLog {
-            item_priority: state
-                .item_precedence
-                .iter()
-                .map(|x| format!("{x:?}"))
-                .collect(),
-            summary: spoiler_summaries,
-            objectives: spoiler_objectives,
-            start_location: SpoilerStartLocation {
-                name: state.start_location.name.clone(),
-                room_id: state.start_location.room_id,
-                node_id: state.start_location.node_id,
-                x: state.start_location.x,
-                y: state.start_location.y,
-            },
-            hub_location_name: hub_room_name,
-            // hub_obtain_route: state.hub_obtain_route.clone(),
-            // hub_return_route: state.hub_return_route.clone(),
-            escape: spoiler_escape,
-            details: spoiler_details,
-            all_items: spoiler_all_items,
-            all_rooms: spoiler_all_rooms,
-            game_data: self.get_spoiler_game_data(),
-        };
 
         let randomization = Randomization {
             objectives: self.objectives.clone(),
@@ -4871,9 +4573,9 @@ impl<'r> Randomizer<'r> {
                 get_spoiler_trail_ids(reverse, best_hub_vertex_id, reverse_cost_idx);
 
             let hub_obtain_route =
-                self.get_spoiler_route(&global, &hub_obtain_trail_ids, forward, false);
+                get_spoiler_route(self, &global, &hub_obtain_trail_ids, forward, false);
             let hub_return_route =
-                self.get_spoiler_route(&global, &hub_return_trail_ids, reverse, true);
+                get_spoiler_route(self, &global, &hub_return_trail_ids, reverse, true);
 
             traverser_pair.forward.pop_step();
             traverser_pair.reverse.pop_step();
@@ -4995,7 +4697,7 @@ impl<'r> Randomizer<'r> {
             details: vec![],
             all_items: vec![],
             all_rooms: spoiler_all_rooms,
-            game_data: self.get_spoiler_game_data(),
+            game_data: get_spoiler_game_data(self),
         };
 
         let randomization = Randomization {
@@ -5069,8 +4771,8 @@ impl<'r> Randomizer<'r> {
         };
         let num_vertices = self.game_data.vertex_isv.keys.len();
         let mut traverser_pair = TraverserPair {
-            forward: Traverser::new(num_vertices, false),
-            reverse: Traverser::new(num_vertices, true),
+            forward: Traverser::new(num_vertices, false, &initial_global_state),
+            reverse: Traverser::new(num_vertices, true, &initial_global_state),
         };
         let start_location_data = self.determine_start_location(
             attempt_num_rando,
@@ -5168,16 +4870,17 @@ impl<'r> Randomizer<'r> {
                 .filter(|x| x.bireachable_traversal.is_some())
                 .count();
             let cnt_flag_bireachable = state
-                    .flag_location_state
-                    .iter()
-                    .filter(|x| x.bireachable_traversal.is_some())
-                    .count();
+                .flag_location_state
+                .iter()
+                .filter(|x| x.bireachable_traversal.is_some())
+                .count();
             info!(
                 "[attempt {attempt_num_rando}] step={0}, bireachable={cnt_bireachable}, reachable={cnt_reachable}, placed={cnt_placed}, collected={cnt_collected}",
                 state.step_num
             );
 
-            let any_progress = cnt_bireachable > last_cnt_bireachable || cnt_flag_bireachable > last_cnt_flag_bireachable;
+            let any_progress = cnt_bireachable > last_cnt_bireachable
+                || cnt_flag_bireachable > last_cnt_flag_bireachable;
 
             if is_early_stop {
                 break;
@@ -5247,822 +4950,6 @@ impl<'r> Randomizer<'r> {
             }
         }
         self.finish(attempt_num_rando, &mut state, &mut rng);
-        self.get_randomization(
-            &state,
-            seed,
-            display_seed,
-            &mut rng,
-            &mut traverser_pair,
-        )
-    }
-}
-
-// Spoiler log ---------------------------------------------------------
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct SpoilerRouteEntry {
-    pub area: String,
-    pub room: String,
-    pub node: String,
-    pub room_id: usize,
-    pub short_room: String,
-    pub from_node_id: usize,
-    pub to_node_id: usize,
-    pub obstacles_bitmask: usize,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub coords: Option<(usize, usize)>,
-    pub strat_name: String,
-    pub strat_id: Option<usize>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub strat_notes: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub energy_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub reserves_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub missiles_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub supers_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub power_bombs_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub relevant_flags: Vec<String>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerLocation {
-    pub area: String,
-    pub room_id: usize,
-    pub room: String,
-    pub node_id: usize,
-    pub node: String,
-    pub coords: (usize, usize),
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerStartLocation {
-    pub name: String,
-    pub room_id: usize,
-    pub node_id: usize,
-    pub x: f32,
-    pub y: f32,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerStartState {
-    max_energy: Capacity,
-    max_reserves: Capacity,
-    max_missiles: Capacity,
-    max_supers: Capacity,
-    max_power_bombs: Capacity,
-    collectible_missiles: Capacity,
-    collectible_supers: Capacity,
-    collectible_power_bombs: Capacity,
-    items: Vec<String>,
-    flags: Vec<String>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerItemDetails {
-    pub item: String,
-    pub location: SpoilerLocation,
-    pub reachable_step: usize,
-    pub difficulty: Option<String>,
-    pub obtain_route: Vec<SpoilerRouteEntry>,
-    pub return_route: Vec<SpoilerRouteEntry>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerFlagDetails {
-    pub flag: String,
-    pub location: SpoilerLocation,
-    pub reachable_step: usize,
-    pub obtain_route: Vec<SpoilerRouteEntry>,
-    pub return_route: Vec<SpoilerRouteEntry>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerDoorDetails {
-    door_type: String,
-    location: SpoilerLocation,
-    obtain_route: Vec<SpoilerRouteEntry>,
-    return_route: Vec<SpoilerRouteEntry>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerLocalState {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub energy_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub reserves_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub missiles_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub supers_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub power_bombs_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub shinecharge_frames_remaining: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cycle_frames: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub farm_baseline_energy_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub farm_baseline_reserves_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub farm_baseline_missiles_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub farm_baseline_supers_used: Option<Capacity>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub farm_baseline_power_bombs_used: Option<Capacity>,
-}
-
-impl SpoilerLocalState {
-    fn new(new_state: LocalState, old_state: LocalState) -> Self {
-        Self {
-            energy_used: if new_state.energy_used == old_state.energy_used {
-                None
-            } else {
-                Some(new_state.energy_used)
-            },
-            reserves_used: if new_state.reserves_used == old_state.reserves_used {
-                None
-            } else {
-                Some(new_state.reserves_used)
-            },
-            missiles_used: if new_state.missiles_used == old_state.missiles_used {
-                None
-            } else {
-                Some(new_state.missiles_used)
-            },
-            supers_used: if new_state.supers_used == old_state.supers_used {
-                None
-            } else {
-                Some(new_state.supers_used)
-            },
-            power_bombs_used: if new_state.power_bombs_used == old_state.power_bombs_used {
-                None
-            } else {
-                Some(new_state.power_bombs_used)
-            },
-            shinecharge_frames_remaining: if new_state.shinecharge_frames_remaining
-                == old_state.shinecharge_frames_remaining
-            {
-                None
-            } else {
-                Some(new_state.shinecharge_frames_remaining)
-            },
-            cycle_frames: if new_state.cycle_frames == old_state.cycle_frames {
-                None
-            } else {
-                Some(new_state.cycle_frames)
-            },
-            farm_baseline_energy_used: if new_state.farm_baseline_energy_used
-                == old_state.farm_baseline_energy_used
-            {
-                None
-            } else {
-                Some(new_state.farm_baseline_energy_used)
-            },
-            farm_baseline_reserves_used: if new_state.farm_baseline_reserves_used
-                == old_state.farm_baseline_reserves_used
-            {
-                None
-            } else {
-                Some(new_state.farm_baseline_reserves_used)
-            },
-            farm_baseline_missiles_used: if new_state.farm_baseline_missiles_used
-                == old_state.farm_baseline_missiles_used
-            {
-                None
-            } else {
-                Some(new_state.farm_baseline_missiles_used)
-            },
-            farm_baseline_supers_used: if new_state.farm_baseline_supers_used
-                == old_state.farm_baseline_supers_used
-            {
-                None
-            } else {
-                Some(new_state.farm_baseline_supers_used)
-            },
-            farm_baseline_power_bombs_used: if new_state.farm_baseline_power_bombs_used
-                == old_state.farm_baseline_power_bombs_used
-            {
-                None
-            } else {
-                Some(new_state.farm_baseline_power_bombs_used)
-            },
-        }
-    }
-}
-
-pub fn get_spoiler_traverse_result(tr: &Traverser) -> SpoilerTraverseResult {
-    let mut out: SpoilerTraverseResult = SpoilerTraverseResult {
-        traversal_number: 0, //tr.traversal_number,
-        prev_trail_ids: vec![],
-        link_idxs: vec![],
-        local_states: vec![],
-        cost: tr.cost.clone(),
-        start_trail_ids: tr.start_trail_ids.clone(),
-    };
-    // for t in &tr.step_trails {
-    //     let old_state = if t.prev_trail_id >= 0 {
-    //         tr.step_trails[t.prev_trail_id as usize].local_state
-    //     } else {
-    //         IMPOSSIBLE_LOCAL_STATE
-    //     };
-    //     let spoiler_local_state = SpoilerLocalState::new(t.local_state, old_state);
-    //     out.prev_trail_ids.push(t.prev_trail_id);
-    //     out.link_idxs.push(t.link_idx);
-    //     out.local_states.push(spoiler_local_state);
-    // }
-    out
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerTraverseResult {
-    pub traversal_number: usize,
-    pub prev_trail_ids: Vec<StepTrailId>,
-    pub link_idxs: Vec<StepTrailId>,
-    pub local_states: Vec<SpoilerLocalState>,
-    pub start_trail_ids: Vec<[StepTrailId; NUM_COST_METRICS]>,
-    pub cost: Vec<[f32; NUM_COST_METRICS]>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerDetails {
-    pub step: usize,
-    pub start_state: SpoilerStartState,
-    pub flags: Vec<SpoilerFlagDetails>,
-    pub doors: Vec<SpoilerDoorDetails>,
-    pub items: Vec<SpoilerItemDetails>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerItemLoc {
-    pub item: String,
-    pub location: SpoilerLocation,
-}
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerRoomLoc {
-    // here temporarily, most likely, since these can be baked into the web UI
-    pub room_id: usize,
-    pub room: String,
-    pub map: Vec<Vec<u8>>,
-    pub map_reachable_step: Vec<Vec<u8>>,
-    pub map_bireachable_step: Vec<Vec<u8>>,
-    pub coords: (usize, usize),
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerItemSummary {
-    pub item: String,
-    pub location: SpoilerLocation,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerFlagSummary {
-    flag: String,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerDoorSummary {
-    door_type: String,
-    location: SpoilerLocation,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerSummary {
-    pub step: usize,
-    pub flags: Vec<SpoilerFlagSummary>,
-    pub doors: Vec<SpoilerDoorSummary>,
-    pub items: Vec<SpoilerItemSummary>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerLink {
-    pub from_vertex_id: VertexId,
-    pub to_vertex_id: VertexId,
-    pub strat_id: Option<usize>,
-    pub strat_name: String,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerRoom {
-    pub room_id: usize,
-    pub name: String,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerNode {
-    pub room_id: usize,
-    pub node_id: usize,
-    pub name: String,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerGameData {
-    rooms: Vec<SpoilerRoom>,
-    nodes: Vec<SpoilerNode>,
-    vertices: Vec<VertexKey>,
-    links: Vec<SpoilerLink>,
-}
-#[derive(Serialize, Deserialize)]
-pub struct SpoilerLog {
-    pub item_priority: Vec<String>,
-    pub summary: Vec<SpoilerSummary>,
-    pub objectives: Vec<String>,
-    pub escape: SpoilerEscape,
-    pub start_location: SpoilerStartLocation,
-    pub hub_location_name: String,
-    // pub hub_obtain_route: Vec<SpoilerRouteEntry>,
-    // pub hub_return_route: Vec<SpoilerRouteEntry>,
-    pub details: Vec<SpoilerDetails>,
-    pub all_items: Vec<SpoilerItemLoc>,
-    pub all_rooms: Vec<SpoilerRoomLoc>,
-    pub game_data: SpoilerGameData,
-}
-
-fn extract_relevant_flags(req: &Requirement, out: &mut Vec<usize>) {
-    match req {
-        Requirement::Flag(flag_id) => {
-            out.push(*flag_id);
-        }
-        Requirement::And(reqs) => {
-            for r in reqs {
-                extract_relevant_flags(r, out);
-            }
-        }
-        Requirement::Or(reqs) => {
-            for r in reqs {
-                extract_relevant_flags(r, out);
-            }
-        }
-        _ => {}
-    }
-}
-
-impl Randomizer<'_> {
-    fn get_vertex_info(&self, vertex_id: usize) -> VertexInfo {
-        let VertexKey {
-            room_id, node_id, ..
-        } = self.game_data.vertex_isv.keys[vertex_id];
-        self.get_vertex_info_by_id(room_id, node_id)
-    }
-    fn get_vertex_info_by_id(&self, room_id: RoomId, node_id: NodeId) -> VertexInfo {
-        let room_ptr = self.game_data.room_ptr_by_id[&room_id];
-        let room_idx = self.game_data.room_idx_by_ptr[&room_ptr];
-        let area = self.map.area[room_idx];
-        let room_coords = self.map.rooms[room_idx];
-        VertexInfo {
-            area_name: self.game_data.area_names[area].clone(),
-            room_name: self.game_data.room_json_map[&room_id]["name"]
-                .as_str()
-                .unwrap()
-                .to_string(),
-            room_id,
-            room_coords,
-            node_name: self.game_data.node_json_map[&(room_id, node_id)]["name"]
-                .as_str()
-                .unwrap()
-                .to_string(),
-            node_id,
-        }
-    }
-
-    fn get_spoiler_start_state(&self, global_state: &GlobalState) -> SpoilerStartState {
-        let mut items: Vec<String> = Vec::new();
-        for i in 0..self.game_data.item_isv.keys.len() {
-            if global_state.inventory.items[i] {
-                items.push(self.game_data.item_isv.keys[i].to_string());
-            }
-        }
-        let mut flags: Vec<String> = Vec::new();
-        for i in 0..self.game_data.flag_isv.keys.len() {
-            if global_state.flags[i] {
-                flags.push(self.game_data.flag_isv.keys[i].to_string());
-            }
-        }
-        SpoilerStartState {
-            max_energy: global_state.inventory.max_energy,
-            max_reserves: global_state.inventory.max_reserves,
-            max_missiles: global_state.inventory.max_missiles,
-            max_supers: global_state.inventory.max_supers,
-            max_power_bombs: global_state.inventory.max_power_bombs,
-            collectible_missiles: global_state.inventory.collectible_missile_packs * 5,
-            collectible_supers: global_state.inventory.collectible_super_packs * 5,
-            collectible_power_bombs: global_state.inventory.collectible_power_bomb_packs * 5,
-            items,
-            flags,
-        }
-    }
-
-    pub fn get_spoiler_route(
-        &self,
-        global_state: &GlobalState,
-        trail_ids: &[StepTrailId],
-        traverser: &Traverser,
-        reverse: bool,
-    ) -> Vec<SpoilerRouteEntry> {
-        let mut route: Vec<SpoilerRouteEntry> = Vec::new();
-
-        if trail_ids.is_empty() {
-            return route;
-        }
-        for &trail_id in trail_ids {
-            let trail = &traverser.step_trails[trail_id as usize];
-            let link_idx = trail.link_idx;
-            let link = self.get_link(link_idx as usize);
-            let new_local_state = trail.local_state;
-            let from_vertex_info = self.get_vertex_info(link.from_vertex_id);
-            let to_vertex_info = self.get_vertex_info(link.to_vertex_id);
-            let VertexKey {
-                obstacle_mask: to_obstacles_mask,
-                ..
-            } = self.game_data.vertex_isv.keys[link.to_vertex_id];
-            let door_coords = self
-                .game_data
-                .node_coords
-                .get(&(to_vertex_info.room_id, to_vertex_info.node_id))
-                .copied();
-            let coords = door_coords.map(|(x, y)| {
-                (
-                    x + to_vertex_info.room_coords.0,
-                    y + to_vertex_info.room_coords.1,
-                )
-            });
-
-            let mut relevant_flag_idxs = vec![];
-            extract_relevant_flags(&link.requirement, &mut relevant_flag_idxs);
-            relevant_flag_idxs.sort();
-            relevant_flag_idxs.dedup();
-            let mut relevant_flags = vec![];
-            for flag_idx in relevant_flag_idxs {
-                let flag_name = self.game_data.flag_isv.keys[flag_idx].clone();
-                if global_state.flags[flag_idx] {
-                    relevant_flags.push(flag_name);
-                }
-            }
-
-            let spoiler_entry = SpoilerRouteEntry {
-                area: to_vertex_info.area_name,
-                short_room: strip_name(&to_vertex_info.room_name),
-                room: to_vertex_info.room_name,
-                node: to_vertex_info.node_name,
-                room_id: to_vertex_info.room_id,
-                from_node_id: from_vertex_info.node_id,
-                to_node_id: to_vertex_info.node_id,
-                strat_id: link.strat_id,
-                obstacles_bitmask: to_obstacles_mask,
-                coords,
-                strat_name: link.strat_name.clone(),
-                strat_notes: link.strat_notes.clone(),
-                energy_used: Some(new_local_state.energy_used),
-                reserves_used: Some(new_local_state.reserves_used),
-                missiles_used: Some(new_local_state.missiles_used),
-                supers_used: Some(new_local_state.supers_used),
-                power_bombs_used: Some(new_local_state.power_bombs_used),
-                relevant_flags,
-            };
-            route.push(spoiler_entry);
-        }
-
-        if reverse {
-            route.reverse();
-        }
-
-        // Remove repeated resource values, to reduce clutter in the spoiler view:
-        for i in (0..(route.len() - 1)).rev() {
-            if route[i + 1].energy_used == route[i].energy_used {
-                route[i + 1].energy_used = None;
-            }
-            if route[i + 1].reserves_used == route[i].reserves_used {
-                route[i + 1].reserves_used = None;
-            }
-            if route[i + 1].missiles_used == route[i].missiles_used {
-                route[i + 1].missiles_used = None;
-            }
-            if route[i + 1].supers_used == route[i].supers_used {
-                route[i + 1].supers_used = None;
-            }
-            if route[i + 1].power_bombs_used == route[i].power_bombs_used {
-                route[i + 1].power_bombs_used = None;
-            }
-        }
-        route
-    }
-
-    fn get_spoiler_route_birectional(
-        &self,
-        state: &RandomizationState,
-        vertex_id: usize,
-        traverser_pair: &TraverserPair,
-    ) -> (Vec<SpoilerRouteEntry>, Vec<SpoilerRouteEntry>) {
-        let forward = &traverser_pair.forward;
-        let reverse = &traverser_pair.reverse;
-        let global_state = &state.global_state;
-        let (forward_cost_idx, reverse_cost_idx) =
-            get_bireachable_idxs(global_state, vertex_id, forward, reverse).unwrap();
-        let forward_trail_ids: Vec<StepTrailId> =
-            get_spoiler_trail_ids(forward, vertex_id, forward_cost_idx);
-        let reverse_trail_ids: Vec<StepTrailId> =
-            get_spoiler_trail_ids(reverse, vertex_id, reverse_cost_idx);
-        let obtain_route = self.get_spoiler_route(global_state, &forward_trail_ids, forward, false);
-        let return_route = self.get_spoiler_route(global_state, &reverse_trail_ids, reverse, true);
-        (obtain_route, return_route)
-    }
-
-    fn get_spoiler_route_one_way(
-        &self,
-        state: &RandomizationState,
-        vertex_id: usize,
-        forward: &Traverser,
-    ) -> Vec<SpoilerRouteEntry> {
-        let global_state = &state.global_state;
-        let forward_cost_idx = get_one_way_reachable_idx(vertex_id, forward).unwrap();
-        let forward_trail_ids: Vec<StepTrailId> =
-            get_spoiler_trail_ids(forward, vertex_id, forward_cost_idx);
-        self.get_spoiler_route(global_state, &forward_trail_ids, forward, false)
-    }
-
-    fn get_spoiler_item_details(
-        &self,
-        state: &RandomizationState,
-        item_vertex_id: usize,
-        item: Item,
-        tier: Option<usize>,
-        item_location_idx: usize,
-        traverser_pair: &TraverserPair,
-    ) -> SpoilerItemDetails {
-        let (obtain_route, return_route) =
-            self.get_spoiler_route_birectional(state, item_vertex_id, traverser_pair);
-        let (room_id, node_id) = self.game_data.item_locations[item_location_idx];
-        let item_vertex_info = self.get_vertex_info_by_id(room_id, node_id);
-        let reachable_traversal = state.item_location_state[item_location_idx]
-            .reachable_traversal
-            .unwrap();
-        SpoilerItemDetails {
-            item: Item::VARIANTS[item as usize].to_string(),
-            location: SpoilerLocation {
-                area: item_vertex_info.area_name,
-                room_id: item_vertex_info.room_id,
-                room: item_vertex_info.room_name,
-                node_id: item_vertex_info.node_id,
-                node: item_vertex_info.node_name,
-                coords: item_vertex_info.room_coords,
-            },
-            reachable_step: traverser_pair.forward.past_steps[reachable_traversal].step_num,
-            difficulty: tier.map(|x| self.difficulty_tiers[x].name.clone()),
-            obtain_route,
-            return_route,
-        }
-    }
-
-    fn get_spoiler_item_summary(
-        &self,
-        _state: &RandomizationState,
-        _item_vertex_id: usize,
-        item: Item,
-        item_location_idx: usize,
-    ) -> SpoilerItemSummary {
-        let (room_id, node_id) = self.game_data.item_locations[item_location_idx];
-        let item_vertex_info = self.get_vertex_info_by_id(room_id, node_id);
-        SpoilerItemSummary {
-            item: Item::VARIANTS[item as usize].to_string(),
-            location: SpoilerLocation {
-                area: item_vertex_info.area_name,
-                room_id: item_vertex_info.room_id,
-                room: item_vertex_info.room_name,
-                node_id: item_vertex_info.node_id,
-                node: item_vertex_info.node_name,
-                coords: item_vertex_info.room_coords,
-            },
-        }
-    }
-
-    pub fn get_spoiler_flag_details(
-        &self,
-        state: &RandomizationState,
-        flag_vertex_id: usize,
-        flag_id: FlagId,
-        flag_idx: usize,
-        traverser_pair: &TraverserPair,
-    ) -> SpoilerFlagDetails {
-        let (obtain_route, return_route) =
-            self.get_spoiler_route_birectional(state, flag_vertex_id, traverser_pair);
-        let flag_vertex_info = self.get_vertex_info(flag_vertex_id);
-        let reachable_traversal = state.flag_location_state[flag_idx]
-            .reachable_traversal
-            .unwrap();
-        SpoilerFlagDetails {
-            flag: self.game_data.flag_isv.keys[flag_id].to_string(),
-            location: SpoilerLocation {
-                area: flag_vertex_info.area_name,
-                room_id: flag_vertex_info.room_id,
-                room: flag_vertex_info.room_name,
-                node_id: flag_vertex_info.node_id,
-                node: flag_vertex_info.node_name,
-                coords: flag_vertex_info.room_coords,
-            },
-            reachable_step: traverser_pair.forward.past_steps[reachable_traversal].step_num,
-            obtain_route,
-            return_route,
-        }
-    }
-
-    pub fn get_spoiler_flag_details_one_way(
-        &self,
-        state: &RandomizationState,
-        flag_vertex_id: usize,
-        flag_id: FlagId,
-        flag_idx: usize,
-        forward: &Traverser,
-    ) -> SpoilerFlagDetails {
-        // This is for a one-way reachable flag, used for f_DefeatedMotherBrain:
-        let obtain_route = self.get_spoiler_route_one_way(state, flag_vertex_id, forward);
-        let flag_vertex_info = self.get_vertex_info(flag_vertex_id);
-        let reachable_traversal = state.flag_location_state[flag_idx]
-            .reachable_traversal
-            .unwrap();
-        SpoilerFlagDetails {
-            flag: self.game_data.flag_isv.keys[flag_id].to_string(),
-            location: SpoilerLocation {
-                area: flag_vertex_info.area_name,
-                room_id: flag_vertex_info.room_id,
-                room: flag_vertex_info.room_name,
-                node_id: flag_vertex_info.node_id,
-                node: flag_vertex_info.node_name,
-                coords: flag_vertex_info.room_coords,
-            },
-            reachable_step: forward.past_steps[reachable_traversal].step_num,
-            obtain_route,
-            return_route: vec![],
-        }
-    }
-
-    fn get_door_type_name(door_type: DoorType) -> String {
-        match door_type {
-            DoorType::Blue => "blue",
-            DoorType::Red => "red",
-            DoorType::Green => "green",
-            DoorType::Yellow => "yellow",
-            DoorType::Gray => "gray",
-            DoorType::Wall => "wall",
-            DoorType::Beam(beam) => match beam {
-                BeamType::Charge => "charge",
-                BeamType::Ice => "ice",
-                BeamType::Wave => "wave",
-                BeamType::Spazer => "spazer",
-                BeamType::Plasma => "plasma",
-            },
-        }
-        .to_string()
-    }
-
-    pub fn get_spoiler_door_details(
-        &self,
-        state: &RandomizationState,
-        unlock_vertex_id: usize,
-        locked_door_idx: usize,
-        traverser_pair: &TraverserPair,
-    ) -> SpoilerDoorDetails {
-        let (obtain_route, return_route) =
-            self.get_spoiler_route_birectional(state, unlock_vertex_id, traverser_pair);
-        let locked_door = &self.locked_door_data.locked_doors[locked_door_idx];
-        let (room_id, node_id) = self.game_data.door_ptr_pair_map[&locked_door.src_ptr_pair];
-        let door_vertex_id = self.game_data.vertex_isv.index_by_key[&VertexKey {
-            room_id,
-            node_id,
-            obstacle_mask: 0,
-            actions: vec![],
-        }];
-        let door_vertex_info = self.get_vertex_info(door_vertex_id);
-        SpoilerDoorDetails {
-            door_type: Self::get_door_type_name(
-                self.locked_door_data.locked_doors[locked_door_idx].door_type,
-            ),
-            location: SpoilerLocation {
-                area: door_vertex_info.area_name,
-                room_id: door_vertex_info.room_id,
-                room: door_vertex_info.room_name,
-                node_id: door_vertex_info.node_id,
-                node: door_vertex_info.node_name,
-                coords: door_vertex_info.room_coords,
-            },
-            obtain_route,
-            return_route,
-        }
-    }
-
-    pub fn get_spoiler_flag_summary(
-        &self,
-        _state: &RandomizationState,
-        _flag_vertex_id: usize,
-        flag_id: FlagId,
-    ) -> SpoilerFlagSummary {
-        SpoilerFlagSummary {
-            flag: self.game_data.flag_isv.keys[flag_id].to_string(),
-        }
-    }
-
-    pub fn get_spoiler_door_summary(
-        &self,
-        _unlock_vertex_id: usize,
-        locked_door_idx: usize,
-    ) -> SpoilerDoorSummary {
-        let locked_door = &self.locked_door_data.locked_doors[locked_door_idx];
-        let (room_id, node_id) = self.game_data.door_ptr_pair_map[&locked_door.src_ptr_pair];
-        let door_vertex_id = self.game_data.vertex_isv.index_by_key[&VertexKey {
-            room_id,
-            node_id,
-            obstacle_mask: 0,
-            actions: vec![],
-        }];
-        let door_vertex_info = self.get_vertex_info(door_vertex_id);
-        SpoilerDoorSummary {
-            door_type: Self::get_door_type_name(
-                self.locked_door_data.locked_doors[locked_door_idx].door_type,
-            ),
-            location: SpoilerLocation {
-                area: door_vertex_info.area_name,
-                room_id: door_vertex_info.room_id,
-                room: door_vertex_info.room_name,
-                node_id: door_vertex_info.node_id,
-                node: door_vertex_info.node_name,
-                coords: door_vertex_info.room_coords,
-            },
-        }
-    }
-
-    pub fn get_spoiler_details(
-        &self,
-        orig_global_state: &GlobalState, // Global state before acquiring new flags
-        state: &RandomizationState,      // State after acquiring new flags but not new items
-        new_state: &RandomizationState,  // State after acquiring new flags and new items
-        spoiler_flag_details: Vec<SpoilerFlagDetails>,
-        spoiler_door_details: Vec<SpoilerDoorDetails>,
-        traverser_pair: &TraverserPair,
-    ) -> SpoilerDetails {
-        let mut items: Vec<SpoilerItemDetails> = Vec::new();
-        for i in 0..self.game_data.item_locations.len() {
-            if let Some(item) = new_state.item_location_state[i].placed_item {
-                if item == Item::Nothing {
-                    continue;
-                }
-                if !state.item_location_state[i].collected
-                    && new_state.item_location_state[i].collected
-                {
-                    let item_vertex_id =
-                        state.item_location_state[i].bireachable_vertex_id.unwrap();
-                    let tier = new_state.item_location_state[i].difficulty_tier;
-                    items.push(self.get_spoiler_item_details(
-                        state,
-                        item_vertex_id,
-                        item,
-                        tier,
-                        i,
-                        traverser_pair,
-                    ));
-                }
-            }
-        }
-        SpoilerDetails {
-            step: state.step_num,
-            start_state: self.get_spoiler_start_state(orig_global_state),
-            items,
-            flags: spoiler_flag_details,
-            doors: spoiler_door_details,
-        }
-    }
-
-    pub fn get_spoiler_summary(
-        &self,
-        _orig_global_state: &GlobalState, // Global state before acquiring new flags
-        state: &RandomizationState,       // State after acquiring new flags but not new items
-        new_state: &RandomizationState,   // State after acquiring new flags and new items
-        spoiler_flag_summaries: Vec<SpoilerFlagSummary>,
-        spoiler_door_summaries: Vec<SpoilerDoorSummary>,
-    ) -> SpoilerSummary {
-        let mut items: Vec<SpoilerItemSummary> = Vec::new();
-        for i in 0..self.game_data.item_locations.len() {
-            if let Some(item) = new_state.item_location_state[i].placed_item {
-                if item == Item::Nothing {
-                    continue;
-                }
-                if !state.item_location_state[i].collected
-                    && new_state.item_location_state[i].collected
-                {
-                    let item_vertex_id =
-                        state.item_location_state[i].bireachable_vertex_id.unwrap();
-                    items.push(self.get_spoiler_item_summary(state, item_vertex_id, item, i));
-                }
-            }
-        }
-        SpoilerSummary {
-            step: state.step_num,
-            items,
-            flags: spoiler_flag_summaries,
-            doors: spoiler_door_summaries,
-        }
+        self.get_randomization(&state, seed, display_seed, &mut rng, &mut traverser_pair)
     }
 }
